+++ pymatgen/io/abinit/abiobjects.py
@@ -159,7 +159,7 @@ def structure_from_abivars(cls=None, *args, **kwargs):
         raise ValueError(f"len(typat) != len(coords):\ntypat: {typat}\ncoords: {coords}")
 
     # Note conversion to int and Fortran --> C indexing
-    typat = np.array(typat, dtype=np.int_)
+    typat = np.array(typat, dtype=int)
     species = [znucl_type[typ - 1] for typ in typat]
 
     return cls(
@@ -244,7 +244,7 @@ or the Virtual Crystal Approximation."""
 
         # [ntypat] list
         znucl_type = [specie.number for specie in types_of_specie]
-        typat = np.zeros(natom, np.int_)
+        typat = np.zeros(natom, int)
         for atm_idx, site in enumerate(structure):
             typat[atm_idx] = types_of_specie.index(site.specie) + 1
     else:
